+++ services/preview/eventgrid/mgmt/2020-04-01-preview/eventgrid/client.go
@@ -1,7 +1,7 @@
-// Package timeseriesinsights implements the Azure ARM Timeseriesinsights service API version 2017-11-15.
+// Package eventgrid implements the Azure ARM Eventgrid service API version 2020-04-01-preview.
 //
-// Time Series Insights client
-package timeseriesinsights
+// Azure EventGrid Management Client
+package eventgrid
 
 // Copyright (c) Microsoft and contributors.  All rights reserved.
 //
@@ -25,11 +25,11 @@ import (
 )
 
 const (
-	// DefaultBaseURI is the default URI used for the service Timeseriesinsights
+	// DefaultBaseURI is the default URI used for the service Eventgrid
 	DefaultBaseURI = "https://management.azure.com"
 )
 
-// BaseClient is the base client for Timeseriesinsights.
+// BaseClient is the base client for Eventgrid.
 type BaseClient struct {
 	autorest.Client
 	BaseURI        string